refactor(dependencies): use separate module for dependencies
