+++ lib/convenient_service/dependencies/extractions/active_support_concern.rb
@@ -3,7 +3,7 @@
 ##
 # NOTE:
 #   Copied from `rails/rails` without any logic modification. Version: v7.0.3.1.
-#   Just wrapped in a namespace `ConvenientService::Extractions::ActiveSupportConcern`.
+#   Just wrapped in a namespace `ConvenientService::Dependencies::Extractions::ActiveSupportConcern`.
 #
 # - https://github.com/marian13/rails/blob/v7.0.3.1/activesupport/lib/active_support/concern.rb
 # - https://github.com/rails/rails/blob/v7.0.3.1/activesupport/lib/active_support/concern.rb
