+++ lib/convenient_service/dependencies/extractions/ruby_middleware.rb
@@ -6,8 +6,8 @@
 #   Just wrapped in a namespace `ConvenientService::Dependencies::Extractions::RubyMiddleware`.
 #   Replaced `require` into `require_relative`.
 #
-# - https://github.com/marian13/ruby-middleware/blob/v0.4.2/middleware.gemspec
-# - https://github.com/Ibsciss/ruby-middleware/blob/v0.4.2/middleware.gemspec
+# - https://github.com/marian13/ruby-middleware/blob/v0.4.2
+# - https://github.com/Ibsciss/ruby-middleware/blob/v0.4.2
 # - https://github.com/Ibsciss/ruby-middleware
 #
 require_relative "ruby_middleware/middleware"
