+++ spec/support/convenient_service/dry.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-ConvenientService.require_assigns_attributes_in_constructor_using_dry_initializer
-ConvenientService.require_has_result_params_validations_using_dry_validation
+ConvenientService::Dependencies.require_assigns_attributes_in_constructor_using_dry_initializer
+ConvenientService::Dependencies.require_has_result_params_validations_using_dry_validation
 
-ConvenientService.require_dry_examples
+ConvenientService::Dependencies.require_dry_examples
