+++ spec/support/convenient_service/rails.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-ConvenientService.require_assigns_attributes_in_constructor_using_active_model_attribute_assignment
-ConvenientService.require_has_attributes_using_active_model_attributes
-ConvenientService.require_has_result_params_validations_using_active_model_validations
+ConvenientService::Dependencies.require_assigns_attributes_in_constructor_using_active_model_attribute_assignment
+ConvenientService::Dependencies.require_has_attributes_using_active_model_attributes
+ConvenientService::Dependencies.require_has_result_params_validations_using_active_model_validations
 
-ConvenientService.require_rails_examples
+ConvenientService::Dependencies.require_rails_examples
